Implement new ADF 'Web' and 'Info' classes to help modularize the code.
