+++ lib/adf_obs.py
@@ -3,18 +3,16 @@ Observations (obs) class for the Atmospheric
 Diagnostics Framework (ADF).
 This class inherits from the AdfConfig class.
 
-Currently this class does four things:
+Currently this class does three things:
 
-1.  Initializes an instance of AdfConfig.
+1.  Initializes an instance of AdfInfo.
 
 2.  Sets the "variable_defaults" ADF variable.
 
-3.  Determines if this is a model vs obs ADF run
-
-4.  If so, then creates a dictionary of what
-    observational dataset is associated with
-    each requested variable, along with any
-    relevant observational meta-data.
+3.  If a model vs obs run, then creates a
+    dictionary of what observational dataset
+    is associated with each requested variable,
+    along with any relevant observational meta-data.
 
 This class also provide methods for extracting
 the observational data and meta-data for use
@@ -26,7 +24,6 @@ in various scripts.
 #++++++++++++++++++++++++++++++
 
 import copy
-import os
 
 from pathlib import Path
 
@@ -37,19 +34,19 @@ from pathlib import Path
 import yaml
 
 #ADF modules:
-from adf_config import AdfConfig
+from adf_info import AdfInfo
 
 #+++++++++++++++++++
 #Define Obs class
 #+++++++++++++++++++
 
-class AdfObs(AdfConfig):
+class AdfObs(AdfInfo):
 
     """
-    Observations class, which reads in
-    config (YAML) files and provides
-    a mechanism to process and retreive
-    relevant config variables.
+    Observations class, which initializes
+    an AdfInfo object and provides
+    additional variables and methods
+    needed for managing observational data.
     """
 
     def __init__(self, config_file, debug=False):
@@ -64,20 +61,14 @@ class AdfObs(AdfConfig):
         #Determine local directory:
         _adf_lib_dir = Path(__file__).parent
 
-        #Read in "basic_info" config dictionary:
-        _basic_info = self.read_config_var('diag_basic_info', required=True)
-
-        #Expand basic info variable strings:
-        self.expand_references(_basic_info)
-
-        #Determin if variable defaults will be used:
-        self.__use_defaults = self.read_config_var('use_defaults', conf_dict=_basic_info)
+        #Determine if variable defaults will be used:
+        self.__use_defaults = self.get_basic_info('use_defaults')
 
         # Check whether user wants to use defaults:
         #-----------------------------------------
         if self.__use_defaults:
             #Determine whether to use adf defaults or custom:
-            _defaults_file = self.read_config_var('custom_defaults', conf_dict=_basic_info)
+            _defaults_file = self.get_basic_info('custom_defaults')
             if _defaults_file is None:
                 _defaults_file = _adf_lib_dir/'adf_variable_defaults.yaml'
 
@@ -87,78 +78,16 @@ class AdfObs(AdfConfig):
         else:
             #Set variable_defaults to empty dictionary:
             self.__variable_defaults = {}
-        #-----------------------------------------
-
-        #Initialize ADF variable list:
-        self.__diag_var_list = self.read_config_var('diag_var_list', required=True)
-
-        #Initialize "compare_obs" variable:
-        self.__compare_obs = self.read_config_var('compare_obs', conf_dict=_basic_info)
-
-        #Initialize "num_procs" variable:
-        #-----------------------------------------
-        temp_num_procs = self.read_config_var('num_procs', conf_dict=_basic_info)
-
-        if not temp_num_procs:
-            #Variable not present, so set to a single processor:
-            self.__num_procs = 1
-        else:
-            #Check if variable is a string and matches keyword:
-            if isinstance(temp_num_procs, str) and \
-               temp_num_procs.strip() == "*":
-
-                #Set number of processors to total number of CPUs
-                #on the node.  Please note that at some point this
-                #may need to be replaced with a DASK implementation
-                #instead:
-
-                #First try to get CPUs allowed by OS for process to use:
-                try:
-                    self.__num_procs = len(os.sched_getaffinity(0))
-                except AttributeError:
-                    #Operating system doesn't support getaffinity, so try
-                    #straight CPU number:
-                    if os.cpu_count():
-                        self.__num_procs = os.cpu_count()
-                    else:
-                        #Something is weird with this Operating System,
-                        #so warn user and then try to run in serial mode:
-                        wmsg = "WARNING!!!! ADF unable to determine how"
-                        wmsg += " many processors are availble on this system,"
-                        wmsg += " so defaulting to a single process/core."
-                        print(wmsg)
-                        self.__num_procs = 1
-                    #End if
-                #End except
-
-            else:
-                #If anything else, then try to convert to integer:
-                try:
-                    self.__num_procs = int(temp_num_procs)
-                except ValueError:
-                    #This variable has been set to something that
-                    #can't be converted into an integer, so warn
-                    #user and then try to run in serial mode:
-                    wmsg = "WARNING!!!!  The 'num_procs' variable"
-                    wmsg += f" has been set to '{temp_num_procs}'"
-                    wmsg += " which cannot be converted to an integer."
-                    wmsg += "\nThe ADF will now default to a single core"
-                    wmsg += " and attempt to run."
-                    print(wmsg)
-                    self.__num_procs = 1
-                #End except
-            #End if
         #End if
-        #Print number of processors being used to debug log (if requested):
-        self.debug_log(f"ADF is running with {self.__num_procs} processors.")
         #-----------------------------------------
 
         #Initialize observations dictionary:
         self.__var_obs_dict = {}
 
         #If this is not a model vs obs run, then stop here:
-        if not self.__compare_obs:
+        if not self.compare_obs:
             return
+        #End if
 
         #Extract the "obs_data_loc" default observational data location:
         obs_data_loc = self.read_config_var("obs_data_loc", conf_dict=_basic_info)
@@ -179,7 +108,7 @@ class AdfObs(AdfConfig):
         #End if
 
         #Loop over variable list:
-        for var in self.__diag_var_list:
+        for var in self.diag_var_list:
 
             #Check if variable is in defaults dictionary:
             if var in _variable_defaults:
@@ -262,6 +191,7 @@ class AdfObs(AdfConfig):
             wmsg += " and check the log for messages.\n"
             wmsg += "!!!!!!!!!!!!!!!\n"
             print(wmsg)
+        #End if
 
     #########
 
@@ -279,20 +209,6 @@ class AdfObs(AdfConfig):
         """Return the '__use_defaults' logical to the user if requested."""
         return self.__use_defaults
 
-    # Create property needed to return "compare_obs" logical to user:
-    @property
-    def compare_obs(self):
-        """Return the "compare_obs" logical to the user if requested."""
-        return self.__compare_obs
-
-    # Create property needed to return "diag_var_list" list to user:
-    @property
-    def diag_var_list(self):
-        """Return a copy of the "diag_var_list" list to the user if requested."""
-        #Note that a copy is needed in order to avoid having a script mistakenly
-        #modify this variable, as it is mutable and thus passed by reference:
-        return copy.copy(self.__diag_var_list)
-
     # Create property needed to return "var_obs_dict" dictionary to user:
     @property
     def var_obs_dict(self):
@@ -300,12 +216,6 @@ class AdfObs(AdfConfig):
         #Note that a copy is needed in order to avoid having a script mistakenly
         #modify this variable, as it is mutable and thus passed by reference:
         return copy.copy(self.__var_obs_dict)
-
-    # Create property needed to return "num_procs" to user:
-    @property
-    def num_procs(self):
-        """Return the "num_procs" logical to the user if requested."""
-        return self.__num_procs
 
 #++++++++++++++++++++
 #End Class definition
